FEAT: Add NavBar component to app
This commit adds a responsive navbar component to allow users to navigate through the page on mobile and desktop devices. SASS and font awesome have been added to the project to aid in styling and a re-usable custom hook created as a utility that can be used to trigger changes based on window re-sizing. Unit tests included as part of TDD to check navbar renders correctly.